breakbread-backend/requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Django==2.2
django-cors-headers==2.5.3
django-filter==2.1.0
django-guardian==1.5.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
Faker==1.0.7
kombu==4.6.3
Markdown==3.1
psycopg2==2.8.2
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
vine==1.3.0
